courseapp/requirements.txt
asgiref==3.8.1
Authlib==1.3.2
babel==2.16.0
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.7
cloudinary==1.42.2
colorama==0.4.6
cryptography==43.0.3
dj3-cloudinary-storage==0.0.6
Django==5.1.6
django-ckeditor==6.7.2
django-cloudinary-storage==0.3.0
django-debug-toolbar==5.0.1
django-js-asset==3.0.1
djangorestframework==3.15.2
drf-yasg==1.21.8
Flask==3.1.0
Flask-Admin==1.6.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-OAuthlib==0.9.6
Flask-SQLAlchemy==3.1.1
google-auth==2.36.0
google-auth-oauthlib==1.2.1
greenlet==3.1.1
humanize==4.11.0
idna==3.10
inflection==0.5.1
inputimeout==1.0.4
itsdangerous==2.2.0
Jinja2==3.1.4
keyboard==0.13.5
MarkupSafe==3.0.2
mysqlclient==2.2.7
numpy==2.1.1
oauthlib==2.1.0
packaging==24.2
pillow==11.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyMySQL==1.1.1
pynput==1.7.7
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
requests==2.32.3
requests-oauthlib==1.1.0
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
Werkzeug==3.1.3
WTForms==3.1.2
